temp

+++ src/main/java/network/ProtocolType.java
@@ -1,9 +1,9 @@
 package network;
 
 public class ProtocolType {
-    static final int REGISTER = 0;
-    static final int MODIFICATION = 1;
-    static final int DELETE = 2;
-    static final int RESPONSE = 3;
-    static final int SEARCH =  4;
+    static final byte REGISTER = 0;
+    static final byte MODIFICATION = 1;
+    static final byte DELETE = 2;
+    static final byte RESPONSE = 3;
+    static final byte SEARCH =  4;
 }
